+++ app/components/modules/Calendar/MainCalendar.tsx
@@ -1,8 +1,8 @@
 'use client'
-import CalendarNavBar from '../molecules/Calendar/CalendarNavBar';
-import Calendar from '../molecules/Calendar/Calendar';
-import styles from '../../styles/MainCalendar.module.css';
-import { CalendarShowType, CalendarShowTypes } from '../../@types/Calendar';
+import CalendarNavBar from '@/modules/Calendar/CalendarNavBar';
+import Calendar from './Calendar';
+import styles from '@/styles/MainCalendar.module.css';
+import { CalendarShowType, CalendarShowTypes } from '@/types/Calendar';
 import React, { useState } from "react";
 const MainCalendar = () => {
