+++ app/components/modules/Calendar/ProviderWrapper.tsx
@@ -1,6 +1,6 @@
 'use client';
 import {SessionProvider} from 'next-auth/react'
-import "../styles/globals.css"
+import ".@/styles/globals.css"
 export default function ProviderWrapper({
   children,
 }: {
