+++ app/components/modules/Minicalendar/DateHelper.ts
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { SHORT_MONTH_NAMES } from '../../../@types/constants';
+import { SHORT_MONTH_NAMES } from '../../../../types/constants';
 
 
 function getDaysOfMonth(input: Date): Date[] {
