+++ app/components/modules/Minicalendar/MiniCalendar.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { CalendarHeader } from './CalendarHeader';
-import {useCalendar} from '../../../hooks/CalendarContext';
+import {useCalendar} from '@/hooks/useCalendar';
 import { CalendarDays } from './CalendarDays';
 import { CalendarYears } from './CalendarYears';
 import React from 'react';
