--- app/components/molecules/SearchBar.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import Input from '../atoms/Input';
-import Button from '../atoms/Button';
-
-const SearchBar = ({styles}) => {
-  return (
-    <>
-      <Input />
-      <Button label="Press me" className="button" />
-    </>
-  );
-};
-
-export default SearchBar;
