+++ app/components/organisms/Sidebar.tsx
@@ -1,7 +1,7 @@
-import CheckboxList from "../molecules/CheckboxList";
-import AvatarMenu from "../molecules/AvatarMenu";
-import MiniCalendar from "../molecules/MiniCalendar/MiniCalendar";
-import styles from '../../styles/Sidebar.module.css';
+import CheckboxList from "@/components/molecules/CheckboxList";
+import AvatarMenu from "@/components/molecules/AvatarMenu";
+import MiniCalendar from "@/modules/Minicalendar/MiniCalendar";
+import styles from '@/styles/Sidebar.module.css';
 const Sidebar = () => {
     return (
         <div className={styles.sidebar}>
