+++ tsconfig.json
@@ -22,7 +22,16 @@
         "name": "next"
       }
     ],
-    "strictNullChecks": false
+    "strictNullChecks": false,
+    "baseUrl": ".",
+    "paths": {
+      "@/types/*": ["types/*"],
+      "@/components/*": ["app/components/*"],
+      "@/hooks/*": ["app/components/hooks/*"],
+      "@/modules/*": ["app/components/modules/*"],
+      "@/Shared/*": ["app/components/Shared/*"],
+      "@/styles/*": ["styles/*"]
+    }
   },
   "include": [
     "next-env.d.ts",
