+++ types/Homepage.ts
